Update patch set 2
Patch Set 2:

From the VNFRQTS meeting on 2017-08-01 there was a request to restructure the directory structure to clearly separate the seed documents from the project deliverables. This new directory structure should now be in place. Unfortunately it deprecates the old directory structure which makes the current push request invalid.  It should be obvious where the open-o docs go in the new directory structure - but if there is a question, please email me. 
please resubmit this push with the document using the revised directory structure.

Patch-set: 2